+++ src/components/global/MpSearchFilter.vue
@@ -92,7 +92,7 @@ onMounted(() => {
     </IconField>
     <ul
       v-if="products.productsInput.length > 0 && inputSearch !== ''"
-      class="w-[14.8rem] md:w-[20rem] absolute text-sm mt-9 md:mt-8 -right-1 md:left-0 font-medium text-gray-900 bg-white border border-gray-200 rounded-lg dark:bg-gray-700 dark:border-gray-600 dark:text-white max-h-[14rem] overflow-y-auto z-10 shadow-lg"
+      class="w-[14rem] md:w-[20rem] absolute text-sm mt-10 md:mt-8 right-0 md:left-0 font-medium text-gray-900 bg-white border border-gray-200 rounded-lg dark:bg-gray-700 dark:border-gray-600 dark:text-white max-h-[14rem] overflow-y-auto z-10 shadow-lg"
     >
       <li
         v-for="(item, index) in products.productsInput"