+++ src/components/global/MpSearchFilter.vue
@@ -13,7 +13,7 @@ const products = useProductsStore()
 const toast = useToast()
 
 const searchProduct = () => {
-  if (inputSearch.value.trim() === '') {
+  if (inputSearch.value || inputSearch.value.trim() === '') {
     showToastSearch('warn', 'Campo vacío', 'Por favor ingrese un valor para buscar.')
     return
   }