+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Change style of checkbox in the admin page.
+- Change size logo in main menu.
 
 ### Fixed
 - Fix the links social media in the footer.
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix shortcut search in mac.
 - Fix remove image preview when the user remove the image in the admin page.
 - Fix checkbox in view page.
+- Fix width of the input search.
 
 ## [1.5.0] - 06/06/2024