+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change style of checkbox in the admin page.
 - Change size logo in main menu.
 - Change input type in search component.
+- Change style button in admin page.
 
 ### Fixed
 - Fix the links social media in the footer.