+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.6.0] - 18/06/2024
 
 ### Changed
 - Change style of checkbox in the admin page.
@@ -149,6 +149,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new version of the project. With products and services.
 
+[1.6.0]: https://github.com/CristhianDaza/megaprom/pull/15
 [1.5.0]: https://github.com/CristhianDaza/megaprom/pull/14
 [1.4.0]: https://github.com/CristhianDaza/megaprom/pull/13
 [1.3.1]: https://github.com/CristhianDaza/megaprom/pull/9